requirements.txt
tensorflow>=2.9.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
python-multipart>=0.0.6
pillow>=9.5.0
numpy>=1.23.5
opencv-python>=4.7.0
matplotlib>=3.7.1
scikit-learn>=1.2.2
pytest>=7.3.1
httpx>=0.24.0
requests>=2.28.2
python-dotenv>=1.0.0
face-recognition>=1.3.0

